% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe-ac-scale.R
\name{ac_scale}
\alias{ac_scale}
\title{Interpretation Scale for Agreement Coefficients}
\usage{
ac_scale(scale = "Landis and Koch")
}
\arguments{
\item{scale}{A string specifying the scale to use for interpretation. The
available scales are: "Landis and Koch", "Fleiss", "Altman", "Cicchetti",
"Koo and Li", and "McHugh". Defaults to "Landis and Koch".}
}
\value{
A tibble containing the lower bound, upper bound, and interpretation
for the selected scale.
}
\description{
Provides interpretation scales for various agreement coefficients based on
different classification schemes. The function supports the following scales:
"Landis and Koch", "Fleiss", "Altman", "Cicchetti", "Koo and Li", and
"McHugh".
}
\examples{
# Example usage with Landis and Koch scale
ac_scale(scale = "Landis and Koch")

# Example usage with Fleiss scale
ac_scale(scale = "Fleiss")

}
\references{
\itemize{
\item \href{https://www.jstor.org/stable/2529310?seq=1#metadata_info_tab_contents}{Landis and Koch (1977)}
\item \href{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.456.3830&rep=rep1&type=pdf}{Fleiss (1981)}
\item \href{https://www.crcpress.com/Practical-Statistics-for-Medical-Research/Altman/p/book/9780412276309}{Altman (1991)}
\item \href{https://www.researchgate.net/publication/232556850_Guidelines_Criteria_and_Rules_of_Thumb_for_Evaluating_Normed_and_Standardized_Assessment_Instrument_in_Psychology}{Cicchetti (1994)}
\item \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3900052/}{McHugh (2012)}
\item \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4913118/}{Koo and Li (2016)}

}
}
