% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac-weights.R
\name{ac_weights}
\alias{ac_weights}
\alias{identity_weights}
\alias{quadratic_weights}
\alias{linear_weights}
\alias{ordinal_weights}
\alias{radical_weights}
\alias{ratio_weights}
\alias{circular_weights}
\alias{bipolar_weights}
\title{Weight-generating functions}
\usage{
ac_weights(categ, weight_type = "unweighted")

identity_weights(categ)

quadratic_weights(categ)

linear_weights(categ)

ordinal_weights(categ)

radical_weights(categ)

ratio_weights(categ)

circular_weights(categ)

bipolar_weights(categ)
}
\arguments{
\item{categ}{A vector containing all categories used. Be careful that the
order matches the the columns and rows of the input data set.}

\item{weight_type}{A character vector only available for the \code{ac_weights}
function. Argument for the generic function to select type of weighting.}
}
\value{
A q × q matrix of weights, where q is the number of categories. The
default argument is "unweighted". With this option, the function will create
a diagonal weight matrix with all diagonal numbers equal to 1, and all
off-diagonal numbers equal to 0. This special weight matrix leads to the
unweighted analysis.
}
\description{
To do the weighted analysis, you may create your own weight matrix, or use
one of the many existing weight-generating functions in the weights.ge.r
script. Each weight function takes single mandatory parameter, which is a
vector containing all categories used in the study. The weight functions
always sort all numeric-type category vectors in ascending order.
Consequently, the weighted coefficients are computed properly only if columns
and rows in the input dataset are ordered the same way. For alphanumeric-type
category vectors, they are assumed to already be ranked following an order
that is meaningful to the researcher.
}
\examples{
ac_weights(c(1:5), "unweighted")
ac_weights(c(1:5), "quadratic")
ac_weights(c("A", "B", "C", "D", "E"), "ordinal")
ac_weights(c("A", "B", "C", "D", "E"), "radical")
ac_weights(c("A", "B", "C", "D", "E"), "ratio")
ac_weights(letters[c(1:5)], "circular")
ac_weights(letters[c(1:5)], "bipolar")
identity_weights(1:5)
quadratic_weights(1:5)
linear_weights(1:5)
ordinal_weights(1:5)
radical_weights(1:5)
ratio_weights(1:5)
circular_weights(1:5)
bipolar_weights(1:5)
}
\references{
\enumerate{
\item Handbook of Inter-Rater Reliability: The Definitive Guide to Measuring
the Extent of Agreement Among Raters. 4th ed. Gaithersburg, MD: Advanced
Analytics.
}
}
