% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-ac_dist.R
\name{calc_ac_dist}
\alias{calc_ac_dist}
\title{Calculate Agreement Coefficients (AC) from a distribution}
\usage{
calc_ac_dist(
  distribution,
  weights = "unweighted",
  categ = NULL,
  conf_lev = 0.95,
  N = Inf,
  test_value = 0,
  alternative = "two.sided",
  show_weights = FALSE
)
}
\arguments{
\item{distribution}{An \emph{nxq} matrix / data frame containing the distribution
of raters by subject and category. Each cell \emph{(i,k)} contains the
number of raters who classified subject \emph{i} into category \emph{k}. An n
x q agreement matrix representing the distribution of raters by subjects (n)
and category (q) (see \code{calc_agree_mat} to convert raw data to distribution).}

\item{weights}{The type of weighting to apply. One of "unweighted",
"quadratic", "linear", "ordinal", "radical", "ratio", "circular",
"bipolar", or a custom matrix of weights.}

\item{categ}{Optional. A vector of category labels. If not provided, the
function will extract unique categories from the data.}

\item{conf_lev}{Confidence level for confidence intervals (default is 0.95).}

\item{N}{Population size for finite population correction (default is Inf).}

\item{test_value}{Hypothesis test value (default is 0).}

\item{alternative}{The type of alternative hypothesis. One of "two.sided",
"less", or "greater".}

\item{show_weights}{Logical whether to show the weights matric with the
results. Default is FALSE.}
}
\value{
A list containing: \item{summary}{A summary of the subjects, raters,
and categories.} \item{ac_table}{A table with calculated agreement
coefficients.} \item{hypothesis}{A string describing the hypothesis test.}
}
\description{
This function calculates agreement coefficients, including Gwet's AC, Fleiss'
kappa, and others, when the input dataset is the distribution of raters by
subject and category. It supports various weight types and hypothesis testing
options.
}
\examples{
library(tidyverse)
library(irrCAC)

rvary2

ex_dist <- calc_agree_mat(data = rvary2,
                          dplyr::starts_with("rater"),
                          subject_id = subject)

ex_dist

calc_ac_dist(distribution = ex_dist)

# Compare to raw
calc_ac_raw(data = rvary2,
            dplyr::starts_with("rater"))

# Compare to irrCAC package
dplyr::bind_rows(
  irrCAC::pa.coeff.dist(ratings = ex_dist),
  irrCAC::bp.coeff.dist(ratings = ex_dist),
  irrCAC::fleiss.kappa.dist(ratings = ex_dist),
  irrCAC::gwet.ac1.dist(ratings = ex_dist),
  irrCAC::krippen.alpha.dist(ratings = ex_dist)
)
}
