% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{calc_p_val}
\alias{calc_p_val}
\title{Calculate P-value from T-statistic}
\usage{
calc_p_val(t, df, alternative = "two.sided")
}
\arguments{
\item{t}{Numeric. The T-statistic value.}

\item{df}{Numeric. Degrees of freedom for the t-distribution.}

\item{alternative}{Character. The alternative hypothesis. One of \code{"two.sided"}, \code{"less"}, or \code{"greater"}. Defaults to \code{"two.sided"}.}
}
\value{
Numeric. The p-value for the specified hypothesis test.
}
\description{
Computes the p-value based on a given T-statistic, degrees of freedom, and hypothesis direction.
}
