% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{calc_test_txt}
\alias{calc_test_txt}
\title{Generate Hypothesis Test Text}
\usage{
calc_test_txt(u, alternative = "two.sided")
}
\arguments{
\item{u}{The hypothesized value for the coefficient in the null hypothesis.}

\item{alternative}{The alternative hypothesis type. Must be one of "two.sided", "less", or "greater". Default is "two.sided".}
}
\value{
A string describing the hypothesis test.
}
\description{
This function generates a descriptive string for a hypothesis test based on the provided test value and the specified alternative hypothesis type.
}
