% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac-weights.R
\name{get_ac_weights}
\alias{get_ac_weights}
\title{Get Agreement Coefficient Weights}
\usage{
get_ac_weights(weights = "unweighted", q)
}
\arguments{
\item{weights}{A character string specifying the type of weight to use. One
of "unweighted", "quadratic", "linear", "ordinal", "radical", "ratio",
"circular", "bipolar", or a custom matrix. Default is "unweighted".}

\item{q}{The number of categories in the agreement matrix.}
}
\value{
A list containing: \item{w_name}{The name of the weight type or
"Custom Weights".} \item{weights_mat}{The corresponding weights matrix.}
}
\description{
This function retrieves or creates a weights matrix for calculating agreement
coefficients based on a specified weighting scheme. Basically, a wrapper
around \code{ac_weights}.
}
\examples{
get_ac_weights(weights = "quadratic", q = 4)
get_ac_weights(weights = matrix(c(1, 0.5, 0, 0,
                                  0.5, 1, 0.5, 0,
                                  0, 0.5, 1, 0.5,
                                  0, 0, 0.5, 1), ncol = 4), q = 4)
}
