% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac-gwet.R
\name{gwet_3_raw}
\alias{gwet_3_raw}
\alias{gwet_ac1_table}
\alias{gwet_3_dist}
\title{Gwet's AC1/AC2 agreement coefficient among multiple raters (2, 3, +)}
\usage{
gwet_3_raw(
  data,
  ...,
  weights = "unweighted",
  categ = NULL,
  conf_lev = 0.95,
  N = Inf,
  test_value = 0,
  alternative = "two.sided"
)

gwet_ac1_table(
  table,
  weights = "unweighted",
  conf_lev = 0.95,
  N = Inf,
  test_value = 0,
  alternative = "two.sided"
)

gwet_3_dist(
  distribution,
  weights = "unweighted",
  categ = NULL,
  conf_lev = 0.95,
  N = Inf,
  test_value = 0,
  alternative = "two.sided"
)
}
\arguments{
\item{data}{A data frame or tibble}

\item{...}{Variable (column) names containing the ratings where each column
represents one rater and each row one subject.}

\item{weights}{is an optional parameter that is either a string variable or a
matrix. The string describes one of the predefined weights and must take
one of the values ("quadratic", "ordinal", "linear", "radical", "ratio",
"circular", "bipolar"). If this parameter is a matrix then it must be a
square matrix qxq where q is the number of possible categories where a
subject can be classified. If some of the q possible categories are not
used, then it is strongly advised to specify the complete list of possible
categories as a vector in parameter \code{categ}. Otherwise, only the categories
reported will be used.}

\item{categ}{An optional parameter representing all categories available to
raters during the experiment. This parameter may be useful if some
categories were not used by any rater in spite of being available to the
raters.}

\item{conf_lev}{The confidence level associated with the agreement
coefficient’s confidence interval. Default is 0.95.}

\item{N}{An optional parameter representing the total number of subjects in
the target subject population. Its default value is infinity, which for all
practical purposes assumes the target subject population to be very large
and will not require any finite-population correction when computing the
standard error.}

\item{test_value}{value to test the estimated AC against. Default is 0.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}

\item{table}{A q×q matrix (or contingency table) showing the distribution
of subjects by rater, where q is the number of categories. This is
the only argument you must specify if you want the unweighted analysis}

\item{distribution}{An \emph{nxq} matrix / data frame containing the distribution
of raters by subject and category. Each cell \emph{(i,k)} contains the
number of raters who classified subject \emph{i} into category \emph{k}. An n
x q agreement matrix representing the distribution of raters by subjects (n)
and category (q) (see \code{calc_agree_mat} to convert raw data to distribution).}
}
\value{
A tbl_df with the coefficient, standard error, lower and upper confidence
limits.

A tbl_df with the coefficient, standard error, lower and upper confidence
limits.

A tbl_df with the coefficient, standard error, lower and upper confidence
limits.
}
\description{
Gwet's AC1/AC2 agreement coefficient among multiple raters (2, 3, +) when the
input data represent the raw ratings reported for each subject and each
rater.

Gwet's AC1/AC2 coefficient and its standard error for 2 raters when input
dataset is a contingency table

Gwet's AC1/Ac2 coefficient (Gwet(2008)) and its standard error for multiple
raters when input dataset is a nxq matrix representing the distribution of
raters by subject and by category. The input data "ratings" is an nxq matrix
showing the number of raters by subject and category. A typical entry
associated with a subject and a category, represents the number of raters who
classified the subject into the specified category. Exclude all subjects that
are not rated by any rater.
}
\examples{
library(tidyverse)

rvary2 <- tibble::tribble(
  ~subject, ~rater1, ~rater2, ~rater3, ~rater4, ~rater5,
  1L,      1L,      2L,      2L,      NA,      2L,
  2L,      1L,      1L,      3L,      3L,      3L,
  3L,      3L,      3L,      3L,      3L,      3L,
  4L,      1L,      1L,      1L,      1L,      3L,
  5L,      1L,      1L,      1L,      3L,      3L,
  6L,      1L,      2L,      2L,      2L,      2L,
  7L,      1L,      1L,      1L,      1L,      1L,
  8L,      2L,      2L,      2L,      2L,      3L,
  9L,      1L,      3L,      NA,      NA,      3L,
  10L,      1L,      1L,      1L,      3L,      3L
)

# More than two raters
gwet_3_raw(data = rvary2,
           dplyr::starts_with("rater"))

# Two raters
gwet_3_raw(data = rvary2,
           rater1:rater2)

# Another example with two raters
rate2 <- tibble::tribble(
  ~rada,     ~radb, ~pop, ~group,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "benign",  12L,     2L,
  "benign",  "Normal",   4L,     5L,
  "benign",  "benign",  17L,     6L,
  "benign", "suspect",   1L,     7L,
  "suspect",  "Normal",   3L,     9L,
  "suspect",  "benign",   9L,    10L,
  "suspect", "suspect",  15L,    11L,
  "suspect",  "cancer",   2L,    12L,
  "cancer",  "cancer",   1L,    16L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "Normal",  21L,     1L,
  "Normal",  "benign",  12L,     2L,
  "Normal",  "benign",  12L,     2L,
  "Normal",  "benign",  12L,     2L,
  "Normal",  "benign",  12L,     2L,
  "Normal",  "benign",  12L,     2L,
  "Normal",  "benign",  12L,     2L,
  "Normal",  "benign",  12L,     2L,
  "Normal",  "benign",  12L,     2L,
  "Normal",  "benign",  12L,     2L,
  "Normal",  "benign",  12L,     2L,
  "Normal",  "benign",  12L,     2L,
  "benign",  "Normal",   4L,     5L,
  "benign",  "Normal",   4L,     5L,
  "benign",  "Normal",   4L,     5L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "benign",  "benign",  17L,     6L,
  "suspect",  "Normal",   3L,     9L,
  "suspect",  "Normal",   3L,     9L,
  "suspect",  "benign",   9L,    10L,
  "suspect",  "benign",   9L,    10L,
  "suspect",  "benign",   9L,    10L,
  "suspect",  "benign",   9L,    10L,
  "suspect",  "benign",   9L,    10L,
  "suspect",  "benign",   9L,    10L,
  "suspect",  "benign",   9L,    10L,
  "suspect",  "benign",   9L,    10L,
  "suspect", "suspect",  15L,    11L,
  "suspect", "suspect",  15L,    11L,
  "suspect", "suspect",  15L,    11L,
  "suspect", "suspect",  15L,    11L,
  "suspect", "suspect",  15L,    11L,
  "suspect", "suspect",  15L,    11L,
  "suspect", "suspect",  15L,    11L,
  "suspect", "suspect",  15L,    11L,
  "suspect", "suspect",  15L,    11L,
  "suspect", "suspect",  15L,    11L,
  "suspect", "suspect",  15L,    11L,
  "suspect", "suspect",  15L,    11L,
  "suspect", "suspect",  15L,    11L,
  "suspect", "suspect",  15L,    11L,
  "suspect",  "cancer",   2L,    12L
)


gwet_3_raw(data = rate2,
           rada, radb)
ratings <- matrix(c(5, 3, 0, 0,
                    3, 11, 4, 0,
                    2, 13, 3, 4,
                    1, 2, 4, 14), ncol = 4, byrow = TRUE)

gwet_ac1_table(table = ratings)

gwet_ac1_table(table = ratings,
               weights = ac_weights(categ = c(1:4),
                                    weight_type = "quadratic"))

my_weights <- matrix(c(1.0000000, 0.8888889, 0.5555556, 0.0000000,
                       0.8888889, 1.0000000, 0.8888889, 0.5555556,
                       0.5555556, 0.8888889, 1.0000000, 0.8888889,
                       0.0000000, 0.5555556, 0.8888889, 1.0000000),
                     ncol = 4, byrow = TRUE)

gwet_ac1_table(table = ratings,
               weights = my_weights)
library(tidyverse)

rvary2 <- tibble::tribble(
            ~subject, ~rater1, ~rater2, ~rater3, ~rater4, ~rater5,
                  1L,      1L,      2L,      2L,      NA,      2L,
                  2L,      1L,      1L,      3L,      3L,      3L,
                  3L,      3L,      3L,      3L,      3L,      3L,
                  4L,      1L,      1L,      1L,      1L,      3L,
                  5L,      1L,      1L,      1L,      3L,      3L,
                  6L,      1L,      2L,      2L,      2L,      2L,
                  7L,      1L,      1L,      1L,      1L,      1L,
                  8L,      2L,      2L,      2L,      2L,      3L,
                  9L,      1L,      3L,      NA,      NA,      3L,
                 10L,      1L,      1L,      1L,      3L,      3L
            )

ex_dist <- calc_agree_mat(data = rvary2,
                          dplyr::starts_with("rater"),
                          subject_id = subject)

ex_dist

gwet_3_dist(distribution = ex_dist)
}
\references{
\enumerate{
\item Handbook of Inter-Rater Reliability: The Definitive Guide to Measuring
the Extent of Agreement Among Raters. 4th ed. Gaithersburg, MD: Advanced
Analytics.
}

Gwet, K. L. (2008). Computing inter-rater reliability and its variance in the
presence of high agreement. British Journal of Mathematical and Statistical
Psychology, 61, 29-48.

Gwet (2008)
\enumerate{
\item Handbook of Inter-Rater Reliability: The Definitive Guide to Measuring
the Extent of Agreement Among Raters. 4th ed. Gaithersburg, MD: Advanced
Analytics.
}

Gwet, K. L. (2014). Handbook of Inter-Rater Reliability. Gaithersburg, MD:
Advanced Analytics, LLC.

\enumerate{
\item Handbook of Inter-Rater Reliability: The Definitive Guide to Measuring
the Extent of Agreement Among Raters. 4th ed. Gaithersburg, MD: Advanced
Analytics.
}

Gwet, K. L. (2008). "Computing inter-rater reliability and its variance in
the presence of high agreement." British Journal of Mathematical and
Statistical Psychology, 61, 29-48.
}
