% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac-krippendorph-alpha.R
\name{krippen_2_table}
\alias{krippen_2_table}
\title{Krippendorff's Alpha}
\usage{
krippen_2_table(
  table,
  weights = "unweighted",
  conf_lev = 0.95,
  N = Inf,
  test_value = 0,
  alternative = "two.sided"
)
}
\arguments{
\item{table}{A qxq matrix (or contingency table) showing the distribution
of subjects by rater, where q is the number of categories. This is
the only argument you must specify if you want the unweighted analysis}

\item{weights}{One of the following to calculate weight based on defined
methods: "unweighted", "quadratic", "linear", "ordinal", "radical",
"ratio", "circular", "bipolar". The default is "unweighted", a diagonal
matrix where all diagonal numbers equal to 1, and all off-diagonal numbers
equal to 0. This special weight matrix leads to the unweighted analysis.
You may specify your own q x q weight matrix here}

\item{conf_lev}{The confidence level associated with the agreement
coefficient's confidence interval. Default is 0.95.}

\item{N}{An optional parameter representing the total number of subjects in
the target subject population. Its default value is infinity, which for all
practical purposes assumes the target subject population to be very large
and will not require any finite-population correction when computing the
standard error.}

\item{test_value}{value to test the estimated AC against. Default is 0.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}
}
\value{
A tbl_df with the coefficient, standard error, lower and upper confidence
limits.
}
\description{
Krippendorff's Alpha and its standard error for 2 raters when input dataset
is a contingency table
}
\examples{
ratings <- matrix(c(5, 3, 0, 0,
                    3, 11, 4, 0,
                    2, 13, 3, 4,
                    1, 2, 4, 14), ncol = 4, byrow = TRUE)

krippen_2_table(table = ratings)

krippen_2_table(table = ratings,
               weights = "quadratic")

krippen_2_table(table = ratings,
               weights = ac_weights(categ = c(1:4),
                                    weight_type = "quadratic"))

my_weights <- matrix(c(1.0000000, 0.8888889, 0.5555556, 0.0000000,
                       0.8888889, 1.0000000, 0.8888889, 0.5555556,
                       0.5555556, 0.8888889, 1.0000000, 0.8888889,
                       0.0000000, 0.5555556, 0.8888889, 1.0000000),
                     ncol = 4, byrow = TRUE)

krippen_2_table(table = ratings,
               weights = my_weights)
}
\references{
Gwet, K. (2014). Handbook of Inter-Rater Reliability: The Definitive Guide to
Measuring the Extent of Agreement Among Multiple Raters, 4th Edition.
Advanced Analytics, LLC;

Krippendorff (1970). "Bivariate agreement coefficients for reliability of
data." Sociological Methodology,2,139-150

Krippendorff (1980). Content analysis: An introduction to its methodology
(2nd ed.), New-bury Park, CA: Sage.
}
