% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagree.R
\docType{package}
\name{lagree-package}
\alias{lagree}
\alias{lagree-package}
\title{lagree:: Calculate various interrater agreement coefficients}
\description{
Calculate various agreement coefficients for 2 or more raters along with
standard errors and confidence intervals. Many packages are available to do
this, but this one returns neat tibbles with results.
}
\examples{
# Example usage:
library(lagree)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emilelatour/lagree}
  \item Report bugs at \url{https://github.com/emilelatour/lagree/issues}
}

}
\author{
\strong{Maintainer}: Emile Latour \email{latour@ohsu.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
