% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagree.R
\docType{package}
\name{lagree}
\alias{lagree}
\alias{lagree-package}
\title{lagree:: Calculate various interrater agreement coefficients}
\description{
Calculate various agreement coefficients for 2 or more raters along with
standard errors and confidence intervals. Many packages are available to do
this, but this one returns neat tibbles with results.
}
\examples{
# Example usage:
library(lagree)

}
