% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-radiologist.R
\docType{data}
\name{radiologist}
\alias{radiologist}
\title{Classification by two radiologists of 85 xeromammograms as normal, benign disease, suspicion of cancer, or cancer}
\format{A tibble with 85 observations (Normal, Benign, Suspect, Cancer) on 2 variables representing
different raters.
\describe{
\item{radiologist_a}{ratings of Radiologist a (see levels above)}
\item{radiologist_b}{ratings of Radiologist b (see levels above)}
}}
\source{
Boyd, N. F., C. Wolfson, M. Moskowitz, T. Carlile, M. Petitclerc, H. A.
Ferri, E. Fishell, A. Gregoire, M. Kiernan, J. D. Longley, I. S. Simor, and
A. B. Miller. 1982. Observer variation in the interpretation of
xeromammograms. Journal of the National Cancer Institute 68: 357–363.

\href{https://www.stata-press.com/data/r13/r.html#r}{Datasets for Stata Base Reference Manual, Release 13}
}
\usage{
radiologist
}
\description{
Classification by two radiologists of 85 xeromammograms as normal, benign
disease, suspicion of cancer, or cancer (a subset of the data from Boyd et
al. (1982) and discussed in the context of kappa in Altman (1991, 403–405).
}
\examples{
radiologist

library(dplyr)
library(tidyr)
radiologist \%>\%
  with(., table(radiologist_a, radiologist_b)) \%>\%
  addmargins()
}
\references{
Altman, D. G. 1991. Practical Statistics for Medical Research. London: Chapman & Hall/CRC
}
\keyword{datasets}
