% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-two-raters.R
\docType{data}
\name{two_raters}
\alias{two_raters}
\title{Generic ratings by two raters}
\format{
A tibble with 125 observations (1, 2) on 2 variables representing
different raters.
\describe{
\item{rater_a}{ratings of rater a (1, 2)}
\item{rater_a}{ratings of rater b (1, 2)}
}
}
\source{
I forget where or why I found this one.
}
\usage{
two_raters
}
\description{
125 subjects rated by two raters
}
\examples{
two_raters

library(dplyr)
library(tidyr)
two_raters \%>\%
  with(., table(rater_a, rater_b))
}
\references{
I forget where or why I found this one.
}
\keyword{datasets}
